Moves block-editor items into own directory.
